organize integration tests
