+++ engine/tests/collection/mod.rs
@@ -1,0 +1,1 @@
+// This tests storage API collection methods.
