+++ engine/tests/database/mod.rs
@@ -1,0 +1,6 @@
+// This tests storage API database methods.
+
+mod create_database;
+mod delete_database;
+mod modify_database;
+mod find_database;
