+++ engine/tests/database/modify_database.rs
@@ -1,0 +1,40 @@
+use engine::{
+    config::Config,
+    Logger,
+    Engine,
+    storage::DB_FILE_EXTENSION,
+};
+use tempfile::tempdir;
+
+#[test]
+fn change_database_description_success() {
+    let db_dir = tempdir().unwrap();
+    let logs_dir = tempdir().unwrap();
+    let config = Config::new(
+        db_dir.path(),
+        logs_dir.path()
+    );
+    let logger = Logger::build(&config);
+    let engine = Engine::build(&config, &logger);
+    let db_name = "test";
+    let description = "Test database";
+    let file_path = db_dir
+        .path()
+        .join(&format!("{}.{}", db_name, DB_FILE_EXTENSION));
+
+    let result = engine
+        .storage_api()
+        .create_database_by_file_path(db_name, &file_path);
+    assert!(result.success);
+
+    let result = engine
+        .storage_api()
+        .change_database_description(&file_path, description);
+    assert!(result.success);
+    assert!(result.data.is_none());
+    assert!(result.error.is_none());
+    assert!(result.log_error.is_none());
+
+    db_dir.close().unwrap();
+    logs_dir.close().unwrap();
+}
