+++ engine/tests/integration.rs
@@ -1,0 +1,6 @@
+// Integration tests are separated by features in different modules.
+// This is the root module of all tests.
+
+mod common;
+mod database;
+mod collection;
